+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master
 
+* [bug] 'shelly config list' doesn't quit with error if cloud has no configuration files
 * [improvement] 'shelly backup download' is now aliased to 'shelly backup get'
 
 ## 0.2.23 / 2013-05-25